NOTICE!!!
classes : OrderController is now CustomerOrderController +
OrderMenuBoundary is now CustomerOrderMenuBoundary
Please check if it changed,if not change it yourself and update
references if needed.

customer menu's create new order leads to a select store window, and
then to a new order menu.
add card is working, view cart is working(With constant data at the
moment).

Client: now has a getArrayList of Store entities from the server. and
can handle getStores from server

Server: can now get an ArrayList of all the stores in the DB and send
them to the Client.

I changed the entities to implement Serializable.

i just found out that in order to send an object of a class we created,
the class must implement Serializable.

Also noted that the .openStream() in the end of the
loader.loader(...............).openStream() allows changes to be made to
the window after the order "load" was called